requirements.txt
streamlit
pillow
easyocr
numpy